Detect and treat local builds specially
In local builds (e.g. osc build with and without KVM), there is no release
version and disturl available, so some checks will fail and raise warnings.
In strict mode those will even cause the build to fail, breaking local builds.

If neither release nor disturl are set in the .containerinfo, treat it as a
local build and skip checks which require them.

Add a new local-proper-derived testcase which is proper-derived just with
disturl and release removed from .containerinfo.